+++ benchmark/compression/module/typebox-system.ts
@@ -1,0 +1,3 @@
+import { TypeSystem } from '@sinclair/typebox/system'
+
+TypeSystem.AllowNaN = true